+++ src/math/quaternion.h
@@ -4,23 +4,25 @@
 #include "vec3d.h"
 
 class Quaternion{
-	
+
 	public:
 
-	double w;
-	double x;
-	double y;
-	double z;
-	
+	float w;
+	float x;
+	float y;
+	float z;
+
 	Quaternion();
-	Quaternion(double w,double x,double y,double z);
-	Quaternion(double w, Vec3D vec);
+	Quaternion(float w,float x,float y,float z);
+	Quaternion(float w, Vec3D vec);
 	void normalize();
 	Quaternion inverse();
 	Quaternion operator*(const Quaternion q);
 	void toUnionNormQuaternion();
-	Vec3D rotation(double angle, Vec3D axis);
+	Vec3D rotation(float angle, Vec3D axis);
 
 };
+
+
 
 #endif
